Split readiness status tests below file size gate
