+++ .pm/tasks/task_d40b6e81581d420481b84a09472b90e1.execution.md
@@ -85,3 +85,12 @@ Example:
 - Actual Result: PR created at https://github.com/eng-cc/oasis7/pull/634.
 - PR Purpose Decision: normal_pr_ci_watch; this is an ordinary implementation/runtime readiness fix, not a manual packaging CI hold.
 - Blocker / Next Action: watch required checks, mergeability, comments, and review threads; if green and only review-approval policy blocks, use authorized admin merge path.
+
+## 2026-06-25 12:20:05 CST / tpm
+- 完成内容: Fixed PR #634 required-gate failure by splitting oversized observability readiness tests into a dedicated test module.
+- 遗留事项: PR checks need to be rerun after push; live ECS package update and rebuild readiness validation remain required after merge.
+- Action: Investigated `required-gate` failure from GitHub Actions run 28146318927 job 83354127995; root cause was `./scripts/check-rust-file-size.sh` reporting `crates/oasis7/src/bin/oasis7_chain_runtime/oasis7_chain_runtime_observability_tests.rs` at 1428 lines, over the 1200-line test-file limit. Moved the genesis/readiness tests to `oasis7_chain_runtime_observability_readiness_tests.rs`, registered the module, and shared existing sample helpers.
+- Validation Command: `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime build_chain_status_payload_`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime observability_transport_tests`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime network_tier_tests::public_testnet_validator`; `cargo fmt --check`; `git diff --check`; `./scripts/check-rust-file-size.sh`; `./scripts/pm/claim-ready.sh --claim-type ready_for_pr --verify-command "env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime build_chain_status_payload_"`
+- Expected Result: CI file-size gate passes and the readiness/status behavior tests remain covered.
+- Actual Result: file-size gate passed with oversized files=0; status payload filter passed 15 tests including the new readiness module; transport tests passed 7; public_testnet network-tier tests passed 7; fmt/diff checks passed; claim-ready passed at 2026-06-25T12:19:57+08:00 with verification_exit_code=0.
+- Blocker / Next Action: commit and push CI fix, then continue PR check watch.
